Implement multiline strings
Currently, RSL doesn't have any multiline string syntax. That changes
with this commit! We opt for a Python/C#/Swift-inpsired mechanism which
makes use of triple double quotes """. I think the implementation I've
gone for is more similar to Swift than anything else. The main different
between RSL string blocks and Python's, which is what I've borrowed from
C# and Swift, is how indentation gets handled.

  text = """
     Hi!
     """

If we print this, we will get (output put between triple dashes ---)

---
Hi!
---

If we now *shift* the ending quotes to the left by one char i.e.

  text = """
     Hi!
    """

we get:

---
 Hi!
---

Notice the extra indent now in front of the 'Hi!'. The reason is that
the amount of whitespace preceding the closing quotes will also be
removed from every line of content in the string. This allows users to
control the intended indentation of their contents by placing their
ending quotes as they like.

Also notice that the 'Hi'! is not preceded by a newline. Content is
assumed to begin *after a newline following the opening quotes*.

The rules are as follows:

- Opening triple quotes can only be followed by a newline or a
  whitespace or a comment.
- Closing quote may only be preceded by whitespace on its line.
- Said closing-quote-whitespace-prefix is removed from every line of
  content in the block.

I am somewhat unsure of the implementation. Prior to this commit, the
lexer has been able to at most peek 2 or so characters ahead of the
current character being consumed. However, with this commit, in order to
implement the dedenting based on the preceding whitespace of the ending
quotes, I opted to perform a regex look-ahead and count said whitespace.
Then, as I proceed through parsing the string, I can instead throw away
the appropriate amount of whitespace at the start of every line.

The alternative implementation involves first parsing the contents, then
going back and removing the whitespace after, but this is really tricky
and would require quite large lexer changes. I don't necessarily see
what's *wrong* with this regex lookup, it just *feels* a little brittle.
But I expect it's still performant, so that's not a concern.

The implementation is a little tricky around how we handle string
interpolation as well, as we treat those as two separte string parsings,
but we need to ensure we don't attempt any whitespace removal from the
beginning of the string "resumption" after the interpolation ends.

Multiline strings aren't 100% complete with this commit, I want to add
two more features:

1) Be able to increase # of quotes in delimiter to e.g. 4 so you don't
   need to escape triple quotes if you want that in your contents. C#
   does this too.
2) Be able to end lines with a \ and have that mean "no newline" from a
   contents perspective. So you can *visually* cut your contents down on
   several lines, but have them actually end up on the same one. Swift
   does this.

Both are fairly low-prio tho, so I will probably move onto other things.