Parser: Allow maps to be defined across lines
Similar to the previous commit where we did this for lists.

Previously, the parser was not written to allow multiline map
definitions. That changes with this commit:

  names = {
    "alice": 1,
    "bob": 2,
  }

- We allow trailing commas.
- There are no requirements or rules around indentation or spacing
  between elements. It's just a stylistic choice.

I actually noticed a mishandling of indendation inside your list/map
definitions in the parser. We should consume not just newlines between
statements, but *any* whitespace, including indents/dedents. The reason
for this is that we parse blocks in whole chunks at a time, usually. So
an if-statement properly handles its own indents, for example. In
between statements, indentation doesn't matter (at least I think, I
haven't thought super deeply about it, will revisit if it doesn't hold
up).