Parser: Allow spreading ternaries across lines
Continuing on in the recent theme of allowing things to be defined
across lines.

This is now allowed (but was not before)

  a = condition
      ? true case
      : false

There aren't really any rules around how you are or are not allowed to
define the ternary with regards to whitespace between characters, as
long as you don't do anything confusing like having a ternary in a block
which then results in inadvertent DEDENT token consumption, such as:

  if true:
    a = "blah" ?
true : false
    print("inside block")
  print("outside block")

We'll consume the DEDENT after the '?', but that DEDENT is expected to
be consumed when we exit the if-block, but won't be there. That's what I
anticipate would happen anyway - I've not stepped through and confirmed.
It does fail, though.

Include some minor touchups on basics.md from yesterday's work on it.