Allow truthy/falsy in logical binary operations
Prior to this commit, this will fail:

  print(2 or 0)

when it really should just print "true".

This was an oversight from when I initially implemented truthy/falsy. I
had not updated the VisitLogicalExpr function that resolves these and/or
ops. Fix is very simple -- we just run the left and right sides through
a Truthy/Falsy transformer method. This commit is mostly adding
comprehensive test coverage.