Add modulo operator
This might be the last major operator that was missing, though I'm still
weighing if integer division "//" should also be added.

Example of what's now possible:

  print(5 % 3)
  a = 10
  a %= 7

outputs:

  2
  3

We also support floats and negative numbers. The latter took a bit of
decision making, since there are two major implementations for how the
behave. The two approaches are referred to as using "truncated" or
"floored" division. RSL joins other major languages in using the former
(C, C++, Java, Swift, Go, C#, etc). The only major languages using the
latter approach are Python and Ruby. The fact that Python uses it does
mean that it wouldn't be unreasonable for us to do the same in RSL, but
I've opted to stay consistent with the majority of the group.

Ultimately, the most important thing is to document the behavior, which
we do.

I also add explicit error throwing for when the denominator is 0, also
for division, which was previously overlooked. It *might* make sense to
add a NaN concept for division, but that can be revisited later.