Improve mixed type equality checking
I hadn't actually tackled this properly before. Previously to this
commit, the following would actually *error* because we were hitting the
'default' switch case of erroring out when two types in a binary
operator were not explicitly handled:

  print("1" == 1)

This commit fixes this up by explicitly checking if an equality check is
being done (== or !=) and if the two types are different, then we return
false.

We update docs around this a bit. I've found the width of the pages to
be quite narrow, so this commit ups it slightly from the default (unsure
what it is) to 60%.

Test 'TestMap_Dot_MixedAssign' gets enabled since that feature was
implemented at some point in the past, but I had forgotten to remove
this test skip.

We also take this opportunity to update the 'array' terminology to
'list' in some places.

This commit still leaves collection types (list, map) comparison
remaining. Will be fixed up in a future commit.