Allow for-each loops through strings
  vowels = 'aeiou'
  for vowel in vowels:
    print('this is a vowel: {vowel}')

This was not allowed before because the 'range' object for for loops
could only be a map or a list. With this change, you can now iterate
through strings, character by character.

It properly handles multibyte characters like emojis and also maintain
attributes like color i.e. the following string would work as you'd
hope:

  a = "hello " + blue("alice") + " 👋"

I was aware of an issue related to len(string) before where it will
actually count bytes instead of distinct runes, but I had just handled
it in that one case where I ran into it. In this change, I again hit it,
so now have introduced a util func StrLen which just wraps the utf8 way
of doing it. This function should always be used when dealing with
strings that contain user input (i.e. RSL code). Otherwise, RSL risks
mishandling strings that contain e.g. emojis.

I've tried find-replacing offenders, but there are probably more.
Something as simple as our len() function actually fell victim to this,
where it'd return an incorrect number.