Add '++' and '--' increment/decrement syntax
I think these are reasonable short-hands to have instead of requiring
+= 1 or -= 1 respectively.

To keep things simple, and avoid any pre-incrment vs. post-increment
confusion, we treat e.g. 'a++' as a *statement* rather than an
*expression*. This is what Go also does (in its fervent pursuit for
simplicity), and I think it's a reasonable simplification, especially
for RSL which doesn't utilize for-i syntax which itself uses explicit
incrementing like a lot of other C-like languages do, which is a big use
case for the syntax treating them as expressions.

We also take this change to shore up the testing and handling of
consecutive pluses and minuses. The lexer particularly was badly behaved
in terms of trying to spit out file headers when it sees mutliple
minuses.