Add & run goimports
This is a tool kinda like go fmt but which handles cleaning go imports.
We add it to our Makefile.

This will be handy because next commit requires me make the code
generation add 'regexp' as a module to every single generated file.
Rather than editing the code generation to only add it where necessary,
we'll simply add this tool to the "generate" step. Doesn't hurt that
it'll keep our imports clean elsewhere, too!