Add 'regex' arg constraint
I'll paste the docs I've written for it, since they're pretty brief:

  If you'd like input strings to match a certain pattern, you can do
  that via a regex constraint.

  args:
    name string
    name regex "[A-Z][a-z]*"

  In this example, a valid name value must start with a capital letter,
  and can then be followed by any number of lowercase letters. No other
  characters will be accepted, so Alice will be a valid value, but bob
  or John123 are not.

  As with other constraints, rad will validate input against this regex,
  and if it doesn't match, it will print an error. The constraint is
  also printed in the script's usage string.

The plan was originally to hold off on this feature until RSL had proper
regex syntax/support, but I haven't been able to decide on what that
should be, and don't think I will (that) soon. I don't want to hold off
on regex constraints any longer since I do think they're valuable and
are a natural addition to the constraints system (it currently just
containing enums). Also, regex support in RSL isn't strictly required as
I thought more about it. With or without it, the string will be a raw
literal with nothing fancy going on.

We compile the regex in the parser, so that an invalid regex can be
detected prior to the interpreter.

Because regexp.Regexp is required to be in one of the generated files,
the template string contains an import for regexp. The previous commit
added goimports to our build step though, so even though the AST code
generation adds this import to every generated file (when only one needs
it), the goimports step which immediately follows it will clean up the
unnecessary import, allowing us to compile.

This commit also includes docs on constraints, covering both enums and
regex constraints.