Docs: Add section on rad block 'color' column modifier
This was a little bit of a journey -- I wanted to show *in the docs* the
coloring that happens here. As best I can tell, there's no good way to
accomplish bespoke coloring in a code block using Material for MkDocs,
unfortunately. You could make some Pygments syntax highlighter, but
that's really overkill and none of the alternatives were better, so I
resorted to a screenshot.

I wanted it to blend in really well with the dark theme for the site so
it doesn't look so overtly like a screenshot. Getting this right was
suprisingly difficult, too, due to color inaccuracies and getting sizing
right, but it works now.

Also, I've become less and less convinced that I need to go the ${} route
for string interpolation. At least, I'm not gonna try it soon, so I fix
up some docs I had written in anticipation of this change, so were
written using the ${} syntax.