Bug fix: Make replace() return an RslString
Just noticed a small class of bugs - functions that haven't been updated
to return RslString, but instead return string. So far (outside of
replace which I'm fixing in this commit), I've identified:

- pick
- upper
- lower
- truncate

There are probably more. These mostly work, but if you do any operations
on the result e.g. concat, then you get an error because our Binary
expression resolver (intentionally) does not support strings (as opposed
to RslStrings). I'm trying to weigh if it's worth coding in a validation
on the returned types, so that it's easier to detect these errors (fail
fast), will think more.