Bug fixes: Ensure native RSL types returned by functions
I was not observing a 'fail fast' philosophy when it came to propagating
returned values through the interpreter in cases where I had incorrectly
implemented built-in functions to return non-native RSL types e.g.
'string' instead of 'RslString'. This has lead to a few bugs that this
commit fixes where you could get away with code working in *some* cases,
but other cases would crash if you did something different with the
returned values.

In this commit, we place an eager validation/check on the return value
of functions which ensures they're of the correct type. Failing this
check is a bug and I should hopefully hit that if I am doing proper unit
testing, and so it will help me avoid this sort of issue going forward.

The functions we fix up are: truncate, lower/upper, pick,
pick_from_resource, now_date.

I also correct something in the docs: JSON doesn't actually
differentiate between ints and floats, but RSL does, so it's not correct
to say we mirror JSON types, exactly.