RLS: Fix shebang completion
Not sure what changed, because sending an empty {} field for
completionProvider worked in the examples I worked on a few days ago,
but I can now get it working again by adding some details into that map.
Specifically, triggerCharacters. Not sure if this is the correct/best
way to go about it, but we've got shebang completion now!