Move textmate bundle into new VSCode extension
Also improve it a bit. The *plan* or *ideal* is to generate this bundle
somehow, so at to avoid having it go out of sync from the our tree
sitter implementation/grammar. Upon digging more into that, it looks
fairly tricky/time consuming, so I'm tabling that for now in favor of
working on RLS some more.