RLS: Utilize RSL tree sitter (RTS) for shebang insertion
Small milestone! With this commit, we now pull in the new RTS library
which is a wrapper on RSL's tree sitter implementation, and leverage it
in order to determine if there's a shebang in the file or not, so we can
offer a code action to insert one.

Since we're dealing with several repos now, this commit also
consolidates the RSL scripts in this repo into a single 'dev.rsl' script
which is capable of building, committing, and releasing for the projects
in this repo. They leverage Makefiles here and here -- I'm wondering if
I should get rid of those too, as they additionally shard out where the
logic lives, which can get a little confusing/unwieldy.