Implement a 'dump' function on RSL trees + build test cases with it
Basically a more detailed 'S Expression' dump. For example, this code:

  a = """
  hello
  friend
  """

gives this dump:

  B: [ 0, 25] PS: [0, 0], PE: [4, 0] source_file
  B: [ 0, 24] PS: [0, 0], PE: [3, 3]   assign
  B: [ 0,  1] PS: [0, 0], PE: [0, 1]     var_path
  B: [ 0,  1] PS: [0, 0], PE: [0, 1]       identifier `a`
  B: [ 2,  3] PS: [0, 2], PE: [0, 3]     = `=`
  B: [ 4, 24] PS: [0, 4], PE: [3, 3]     primary_expr
  B: [ 4, 24] PS: [0, 4], PE: [3, 3]       literal
  B: [ 4, 24] PS: [0, 4], PE: [3, 3]         string
  B: [ 4,  8] PS: [0, 4], PE: [1, 0]           string_start `"""\n`
  B: [ 8, 20] PS: [1, 0], PE: [2, 6]           string_contents
  B: [ 8, 13] PS: [1, 0], PE: [1, 5]             string_content `hello`
  B: [13, 14] PS: [1, 5], PE: [2, 0]             string_content `\n`
  B: [14, 20] PS: [2, 0], PE: [2, 6]             string_content `friend`
  B: [20, 24] PS: [2, 6], PE: [3, 3]           string_end `\n"""`

"B" shows the Byte [start, end], "PS" is point start [row, col], and
ditto for "PE" but for point end.

We then print the node tree, with indentation according to tree levels.
For leaf nodes, we print their contents, which should be accurate to the
final "value" of the node.

To line up the B, PS, PE numbers, we do a pre-walk of the tree to find
the largest numbers, so we can pad them all and keep the aligned when
we do the final print.

These dumps are really a quite thorough 'state' of a produced tree
sitter tree, so taking inspiration from how tree sitter itself does
tests, I've also made a quick test harness which compares before/after
of these dumps (see test/dumps). This will probably be the best way to
test the tree going forward.