Revamp: Add map support, support non-string keys
Previously, only strings were allowed as keys. That's because this is a
requirement of JSON, but it's a bit too limiting on a programming
language. In this re-implementation of maps, we allow non-collection
types e.g. ints, bools, etc (but no lists/maps) as keys.

This commit includes dot syntax, map lookup, and map modification
syntax.