Revamp: Restructure function framework
In this commit, we use this 'revamp' sprint as an opportunity to also
put in place some infrastructure for defining, validating, and executing
functions.

We define a struct which represents callable functions in RSL, including
metadata like named args, expected types, etc, and we use this to
validate arguments invoked at runtime before calling the actual
function. This means we don't need to do all this generic validation for
each function, like providing the correct number of args, etc.