Revamp: Implement time function 'now()' + 'type_of()'
Previously, we had the following functions:

  now_date
  now_year
  now_month
  now_day
  now_hour
  now_minute
  now_second
  epoch_seconds
  epoch_millis
  epoch_nanos

These have now been replaced with a single function 'now()' which
returns a map containing *each* of these values. Each value is then
conventiently accessible using the map doy syntax, for example:

  time = now()
  time.date
  time.month
  time.hour
  time.epoch.millis

This does have the downside of eagerly computing all the values. It
might make sense to, in future, allow specifying which specific value
you're after e.g. now(["hour", "minute"]) and it will return a map with
just those values. But I think we'll probably be fine, this stuff is
quick.

We also add another function 'type_of()'. If you give this an argument,
it will return the type of it as a string. This could be handy in some
situations, but it's also handy for my unit tests as I can use it to
verify (in the test RSL script itself) what types things are.