Revamp: Implement http functions
Another slight rework. We add http_METHOD -style functions for each of
the 9 http methods, and make them all have the same signature i.e. only
the url is mandatory. Body and headers are named args. They all
implemented using the same bit of code, so behave the same, just with
altered HTTP methods when executed.

Headers can be string->string or string->list[string], as is possible
with the HTTP spec.

I was considering having a single 'http' function which would take the
method as a string, but I am quite interested in the idea of UFCS i.e.
the ability for these to be equivalent:

  my_func(arg1, arg2)
  arg1.my_func(arg2)

In a UFCS-forward world, I'd want it to look like:

  url.http("get")

But this would mean a non-UFCS call would look like

  http(url, "get")

which is very unusual for an HTTP request function - languages typically
have the method first, which I agree with being more intuitive.

I might *still* implement an 'http' function down the road so that
there's an approach which lets the method be a string variable, but I
leave it out for now. Each method having its own function means that
both the UFCS and non-UFCS calls look reasonable:

  http_get(url, body=body)
  url.http_get(body=body)