Revamp: The Great Deletion™ (delete handwritten lexer/parser, etc)
Huge milestone! With recent commits, we're (in theory*) at feature
parity with "old rad" which used the handwritten lexer/parser, but we
can now do all the same things with our tree-sitter-powered
lexer/parser/interpreter. It's been a huge lift but it positions rad a
lot better going forward.

This commit deletes most (if not all) the things we no longer need.
There will be follow-up work to ensure all the tests are passing and
that we're at feature parity from a unit-test perspective as well, but
no doubt a lot of them are failing for superfluous reasons atm e.g.
differently formatted lists, error messages, etc.