Revamp: Get remainder unit tests working
A few misc changes to get remaining unit tests passing. Most are
superfluous failures e.g. the error msg has changed, or our way of
printing strings/lists have changed, but there are *some* behavior
changes, such as:

- Allow single-stmt defer statements
- Fix 'in' expression for strings in lists + others
- Add validation on not giving *too many* args to functions
- Fix sorting characters in strings
- Fix color functions
- Allow escaping interpolations in strings ( "\{...")
- Fix string formatting in interpolation
- Improve validation around mods on undefined fields in rad blocks
- Add key/list length equivalence validation for pick functions
- Fix matching logic for pick functions
- Fix logic around slicing when the start index is greater than the end
  index.
- Improve 'contains' behavior on lists

There are several tests that I opt to t.Skip for now because they're a
bit more involved to fix. I've made note of the required work to get
them working, and will be prioritizing them appropriately.

We also introduce an "IsTest" global bool to control some of our
panic-catching behavior when unit testing. It's not ideal to use because
it cuts out what we test in our unit tests, but the alternative of
working around some test-plumbing to get this working as desired is
perhaps more undesirable. We'll try to keep our usage of IsTest to a
minimum.