Strings: Do not escape delimiter alternatives in contents
Currently, if I have a string

  "blah \' blah"

It will become

  blah ' blah

but it should be

  blah \' blah

The escaping logic on delimiters should only apply when the *current
string's delimiter* is being escaped. Otherwise, they should be left
as-is.

This commit accomplishes that by tracking delimiters as a stack in the
interpreter. Not the cleanest, and is akin to what RSL's scanner.c does
in the tree sitter implementation, but it does the job. The alternative
is we fix this in the scanner.c/grammar itself, which I might, when I go
look at some other issues with that area atm.