Implement global flags: --SRC, --RSL-TREE
--SRC is mostly for convenience, so you can easily tag it on and see the
actual code you're running, especially when you're debugging something.

--RSL-TREE is a cleaner way to look at the syntax tree, than the current
approach of --RAD-DEBUG, which prints a bunch of other things and also
runs the script itself. RSL-TREE simply prints the syntax tree and does
not run the script.

These flags, together with --VERSION, may run together, but cause rad to
exit before running the script. This is just what I think is preferred
behavior -- being able to tag these flags onto the end of a command
which otherwise runs the script, and instead have them shortcircuit
running the script, makes them more ergonomic for debugging. Might
change my mind on this in the future.