Add built-in command 'new'
This is a new concept, so this commit sets up a bit of a framework
around it. It's currently only built for building in commands that are
really RSL scripts under-the-hood, shipping with the rad CLI tool.

This uses Go embedding, which allows including RSL code in the binary,
and which can be loaded at runtime.

This 'new' command simply sets up a target file with some
templating/execution permissions, saving some time when making new
scripts.

We change the global 'rad -h' usage string a bit to include commands,
and I also modify it to link to GitHub/docs and provide some extra basic
tips.

I also move some shared util methods into the 'com' common package.