Add 'confirm' modifier to shell commands
A common workflow I've started doing is something like this:

  cmd = `...`
  if confirm("Run cmd {cmd}? [y/n] > "):
    $!cmd
  else:
    print("Aborting!")
    exit(1)

It'd be nice to have a quick way to express this. With this commit, you
can now added a 'confirm' modifier before the shell execution like so:

  confirm $!`...`

Which is essentially equivalent to the aforementioned manual
if-statement. rad will print the command about to be executed, and then
on a new line, it will request confirmation (similar to confirm()), and
if said confirmation is rejected, the script exits with a non-0 error
code.