Implement if clauses with pre statements
Heavily inspired by Go, I've found this construct quite handy:

  if a = input(); len(a) > 10:
    print("too long!")

before running a conditional for an if clause, you can have this
statement (often involving an assignment) which runs before the
condition, allowing you to use that assignment in the condition as well.

This is a bit more concise than alternatives.