Bump tree sitter parser - fix confirm() mis-parse issue
Pasted explanation from tree-sitter-rsl:

Basically, since adding 'confirm' as a modifier on shell commands,
confirm() function calls have been failing to get parsed correctly.

This is because tree sitter prioritizes parsing keywords over
identifiers. So "confirm" being a keyword in the context of shell
commands, is enough for "confirm" to get tokenized as its own thing,
rather than an identifier, even in invalid contexts. This means
confirm() calls were incorrectly parsed.

My *fix* (which I *hope* is not gonna backfire in some way) is to write
explicit aliases for the call grammar rule which basically says
"confirm" as a keyword should *look* like an identifier. This allows
tree sitter to continue parsing the call rule, and doing it correctly.

I add aliases for all 3 of the keywords that can prefix shell commands
(confirm, unsafe, quiet), as they're the only ones I'm aware of right
now that can cause issues. As more keywords get added here, we'll need
to keep adding aliases (unfortunately). I've left a couple of commends
in the grammar to explain this and hopefully make it easier to remember
to do.