Allow list comprehension exprs to return 0 results
  [<expr> for <identifier> in <collection>]

Previously, we required 'expr' to return at least one value, and we'd
always append the first one. This is potentially undesirable in two
scenarios:

1) If a function returns several values, only the first one becoming
   part of the resulting list seems arbitrary and may be undesirable,
   depending on what the user is after.
2) More significantly, it means you can't do list comprehensions with
   expressions that return nothing, for example with 'void' functions
   like 'print'.

  [print(n) for n in numbers]

This is now supported, and is basically a concise for loop.
One difference from Python is that Python will return a list of Nones in
this case, as each print returns None. I am hoping to avoid introducing
a null type in RSL, so we instead return an empty list in this case.

Basically, the result of a list comprehension is a list containing any
returned values from each eval of the expression, no matter how many
results it contains.

I don't *think* this will be undesirable for anyone. You could argue
that, instead of doing this sorta print via a list comprehension, you
should do it via a one liner for loop:

  for n in numbers: print(n)

I opt to *still* allow the list comprehension for three reasons:

1) Python allows it, so users may expect it to be legal (and I don't
   have any objection to it).
2) It may seem a little inconsistent that only *some* functions are
   allowed in a list comprehension. Not as convincing an argument, but
   it's there.
3) It arguably lessens the cognitive load a *little* bit, as we don't
   require users to think about how they need to write their one-liner
   for loop. If they can just *always* do a list comprehension-style
   construct, then they don't need to think "oh wait my expr doesn't
   return anything so I can't do a list comprehension".

This commit also un-skips some tests that are now working, after recent
tree sitter and interpreter work.