Implement switch stmts (revamp)
The revamp left these out because I wanted some time to think them
through a little. I've still not made up my mind completely on how they
should work. The part I'm unsure about is the implicit 'choice' aspect I
had implemented before the revamp, i.e. where cases could be selected
based on how their values peformed string interpolation, and we'd pick
the case that used the most valid (and no invalid) variables in its
interpolation.

It's a cool idea, but I worry that it's a bit complicated and confusing
to users, and the value of them is not quite clear to me. I'll mulling
on it.

But for now, it's clear we need *some* vanilla switch construct. That's
what this commit adds to the interpreter (given the upstream tree sitter
grammar changes).