VSC Extension: Detect RSL language on rad shebang, update TM bundle
TM Bundle now considers e.g. switch, confirm, and/or to be keywords and
highlights them. I moved the function regex ahead of the global keywords
tho, so that confirm (for example) appears like a function instead of
keyword if it's followed by ().

I also include a new dev script for the extension, to make dealing with
version bumping, reinstalling for dev, and releasing with vsce easier.