Only keep first return value from expr in list comprehension
I had recently implemented it such that we append *all* the return
values, but the issue I ran into was that this:

  [parse_int(s) for s in strings]

would return something like this:

  [ 1, { }, 2, { } ]

because parse_int, unless one result was explicitly requested from it,
would return an empty 'error' object after each 'parsed' return value.

We could address this function-by-function by e.g. implementing
parse_int to only return a single value if no return constraints are put
on it, but I think there are a couple of issues with this:

1. Would need to remember to do this in the implementation in each
   function.
2. For every function, we'd need to make a decision on which return
   values are worth returning in this case. That might not always be
   clear.

So instead, for consistency, we will always just take the first result
(if it exists) and append it to the resulting list. This might be
undesirable in some functions where you want both return values, but for
these specific cases, the user can resort to something other than a list
comprehension i.e. a manual for-loop-append structure.