Validate script for errors before attempting script arg parse
Prior to this commit, we try to parse script-defined args prior to
checking the script's syntax tree for invalid nodes. That doesn't really
make sense - we risk triggering an error related to arg parsing, but
said error really is related to the fact that the script has invalid arg
syntax, which tree sitter found, but we haven't checked yet, so we
report the wrong thing.

In this commit, we move the validation up to right after we've handled
global flags (which shouldn't be script-dependent).