Misc: Adjust standard streams output for usage strings
Previously, I was always printing -h usage strings to stderr, I didn't
think too much of it. The main reason was that it would be beneficial
for when --SHELL was enabled, and we wouldn't want the encapsulating
bash script to eval the usage string.

However, discussion on #2 got me looking into this more, and for all
other (non-error) purposes, it's actually more conventional to print
this stuff to stdout, so we'll do that. Only if there's an error or if
we've got --SHELL enabled, will we print usage to stderr.

While here, we also fix up script-less invocations with unknown global
flags. Before this commit, we'd just ignore those, but we'll now
re-enable the 'throw on unknown flags' behavior and re-parse args when
we see we don't have a script (as appears to be the easiest workaround
given some pflag limitations around inspecting unknown args).