Add 'zip' function
Pretty standard way to combine lists into a list of lists by index.

This function might be particularly handy in RSL when used with rad
blocks, as rad blocks populate parallel lists, and so there can be many
use cases where you then want to for-loop through these parallel arrays.

Zip will aid in that.

This will become particularly handy in a future addition to for loops:

  for idx, valA, valB, valC in zip(listA, listB, listC):
    ...

i.e. the ability to easily extract values from a zipped list together.