Allow list-of-list unpacking in for loops
Change for loops such that, if 3 or more identifiers are defined on the
left, while the first will be the index, the 2nd onwards will be defined
as the 'unpacked' values on the right, expecting it to be a list of
lists. To demonstrate:

  a = [10, 20, 30]
  b = ['a', 'b', 'c']

  for idx, val_a, val_b in zip(a, b):
    print(idx, val_a, val_b)

will output

  0 10 a
  1 20 b
  2 30 c

One wrinkle is list comprehensions. So far (and with this commit), we
have made the for-clause equivalent between for-each loops and list
comprehensions, so they behave the exact same way. However, it does mean
list comprehensions with unpacking look like this:

  out = [a * b for _, a, b in zip(list_a, list_b)]

where _ is used as a placeholder for the index, and indicates that it's
not used.

Ultimately, my main concern is that this would catch people out,
especially if they're expecting Python-like behavior. So they might
write e.g.

  for a, b in ...

and expect 'a' to be the first value, but it's really the index.

One alternative is that we make for-clauses slightly different for list
comprehensions and don't offer getting out the index. This would reflect
the assumption (that I'd generally agree with) that the index is rarely
useful in the context of list comprehensions.

Ultimately, I've decided to not do anything at this point, and am
keeping the for-clause consistent. I've posted on
/r/ProgrammingLanguages here, requesting input, will see what comes of
it:

 https://www.reddit.com/r/ProgrammingLanguages/comments/1j1hirv/looking_for_input_on_for_loops/