Allow script args to shadow global flags
As discussed on #2, it'd be appealing to make the global flags (which
are currently all caps) lowercase. The reason they were made uppercase
was to avoid collisions with script args users might want to declare. I
hadn't made my mind up for whether we should just shadow global flags in
that case, or stick to making them unlikely to be used (e.g. all caps),
but then prevent shadowing, so that they're always available.

Thinking more about this, we should lowercase global flags, and allow
shadowing them with script args, if there's a conflict.

This commit implements the shadowing behavior, but the global flags
remain all caps - we'll change that in an separate commit.

This commit also removes the --STDIN flag. A more common convention that
you see in CLIs is to instead allow '-' as the first arg, indicating
"read from stdin". Vim does this, visual studio code's cli 'code' does
this, etc, so we'll follow that convention and simplify the logic for
our stdin reading decision. A purpose of this flag though was to give
the script a name, given that's usually taken from the script file's
name, but with stdin, we can't. Without the flag, we really cannot. So
right now, we leave the name empty, but I may add a global flag which is
something like "--script-name" which basically tells rad what to use.
Bit of an edge case, at this point though.

The change in stdin logic makes implementing the shadowing rules easier.
I've changed the rad runner code to basically try to speed-run reading
the RSL code if given ASAP, so that we can look at its flags, know which
global flags to shadow, and then initialize all our global
(flag-dependent) state such as printing, debugging, etc.