Make all global flags lowercase
Plus a slight adjustment on when the full flag name is shadowed. Before,
if the global full flag was shadowed, we'd put the shorthand in as the
full name, so both e.g. -d and --d were valid for e.g. "debug". However,
that's a little weird and the --d is really not that helpful. If the
full name is shadowed, but the shorthand is fine, let's just allow the
shorthand.

We also make a note of this shadowing logic in the docs.