Hide some global flags when --help invoked on script
If --help is invoked a script, the user is likely interested in usage
specific to that script. Global flags are good to include, particularly
if they're useful *when invoking a script*, but if we print *all* the
global flags, then there's a lot of noise, and we risk overloading the
user with too much information.

Let's hide some flags, either because

1) they're not really relevant when seeking help on a particular script
2) the global flag is quite niche, and not worth always printing

For this commit, we hide the below flags (justification according to the
two above bullet points included):

- rad-debug     (2)
- shell         (2)
- version       (1)
- rsl-tree      (2)
- mock-response (2)

Not always printing all flags is potentially a little divisive, but
this "too much info" problem risks only getting worse, as we add more
global flags.