Global --src flag: print script path header
Often, I want to see the path of a script. And sometimes the code.
Sometimes both!

Rather than having to switch between doing 'which myscript' and
'myscript --src', let's just make --src print the location as well, in a
clear header.

We're careful to not wreck redirects of the source, by respecting IsTty
and printing the header to stderr.