Add function 'find_paths'
You'll often want to iterate through files in a particular folder. Right
now, to do so, you have to either hard-code a list of paths, or utilize
shell commands and parse the output.

Let's support this natively with a new function, instead.

For the above example, you can do something like this:

  for path in find_paths(dir, depth=depth, relative="cwd"):
    file = get_path(path)
    print(file.base_name, file.type)

Perhaps a liiiittle tricky that you need to remember that find_paths
defaults to a 'relative' mode of 'target', but that's how e.g. 'ls'
works. Need to think more about if this default makes sense.